Correcciones del 6/06
Se realizan las correcciones propuestas por el instructor el día 6 de junio
