+++ src/main/resources/application.properties
@@ -10,3 +10,5 @@ spring.flyway.enabled=true
 spring.flyway.locations=classpath:/db/migration
 
 spring.mvc.pathmatch.matching-strategy=ANT_PATH_MATCHER
+
+api.userClient.url=https://g2-users-ms.herokuapp.com/api/users-ms/users/
